Move camera around.

+++ sandbox/source/game.cpp
@@ -12,7 +12,7 @@ Game::Game(const lich::Console_Args &console_args) :
 {
 	push_layer<Events_Logger_Layer>();
 	push_overlay<lich::Imgui_Layer>(_window->handle());
-	push_layer<Render_Layer>(app_spec().width / app_spec().height);
+	push_layer<Render_Layer>((float)app_spec().width / (float)app_spec().height);
 		
 	_logger.trace("Hello!");
 }
